LS-1538 Problem Saving Changes modal error when attempting to update an email address.
This PR adds logic to clear confusion for a user who registers under a different email and attempt to update it to one that already exists in Gigya. Also adds M2 db rollback logic for this scenario when Gigya fails to update to prevent customer data from being out of sync between M2/Gigya.